requirements.txt
slack-sdk==3.26.1
slack-bolt==1.18.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
pinecone-client==2.2.4
python-dotenv==1.0.0
tiktoken==0.5.2